ai-review-service/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
motor
redis
python-dotenv
pydantic-settings
